Failing attempt at implementing Polymer3/X
It seemed like a good idea at the time. Implement
a library that was almost purely JS, with proper
webpack building. That was a mistake. LasaleFamine's
PolymerX doesn't even build in the state you get it
from yo, nonetheless with a project built on top of it.
I was very much looking forward to proper JS includes
but that will have to be a future implementation.
I'm half tempted at this juncture to write my own
custom objects library. I think that Polymer 3 could
be so much further along if it didn't have to support <
Polymer 2.
I went through so many iterations on this I don't know where
to begin. Currently it's failing on looking for a sw.js file
in dev, which will never exist per this config.

I have run into so many problems with this build, ranging from
unmet and seemingly unresolveable peer dependencies to monkeypatching
webpack itself. I advise not to use this package branch, except as
reference for what not to do, and maybe some components could
be abstracted, such as the track component.